Move utils function to a modules folder
