+++ react/modules/ModalManager.ts
